+++ ksa_compliance/output_models/prepayment_invoice/prepayment_invoice_factory.py
@@ -7,7 +7,7 @@ from .invoice_line_factory import invoice_line_create
 from .models import PrepaymentInvoice
 
 
-def prepayment_invoice_factory_create(zatca_fields_dto: dict, doc: SalesInvoice) -> PrepaymentInvoice:
+def prepayment_invoice_factory_create(doc: SalesInvoice) -> PrepaymentInvoice:
     prepayment_invoice = PrepaymentInvoice(
         prepaid_amount=_get_prepaid_amount(doc),
         currency=doc.currency,
@@ -19,10 +19,11 @@ def prepayment_invoice_factory_create(zatca_fields_dto: dict, doc: SalesInvoice)
 def _get_prepaid_amount(doc: SalesInvoice) -> float:
     prepaid_amount = 0
     for row in doc.advances:
-        tax_amount = frappe.db.get_value(
-            row.reference_type,
-            row.reference_name,
-            'total_taxes_and_charges',
-        )
-        prepaid_amount += row.allocated_amount + tax_amount
+        if frappe.db.get_value('Payment Entry', row.reference_name, 'custom_prepayment_invoice'):
+            tax_amount = frappe.db.get_value(
+                row.reference_type,
+                row.reference_name,
+                'total_taxes_and_charges',
+            )
+            prepaid_amount += row.allocated_amount + tax_amount
     return abs(prepaid_amount) if prepaid_amount else 0.0